training/requirements.txt
torch==2.6.0
torchvision
torchaudio
transformers>=4.47.0
peft>=0.14.0
trl>=0.13.0
datasets>=3.0.0
accelerate>=1.3.0
bitsandbytes>=0.45.0
pyyaml>=6.0
